+++ src/ftxui/dom/canvas.cpp
@@ -410,11 +410,11 @@ void Canvas::DrawPointEllipseFilled(int x1,
     e2 = 2 * err;
     if (e2 >= dx) {
       x++;
-      err += dx += 2 * static_cast<long>(r2) * r2;  // NOLINT
+      err += dx += 2 * r2 * r2;
     }
     if (e2 <= dy) {
       y++;
-      err += dy += 2 * static_cast<long>(r1) * r1;  // NOLINT
+      err += dy += 2 * r1 * r1;
     }
   } while (x <= 0);
 
@@ -470,7 +470,7 @@ void Canvas::DrawBlockOn(int x, int y) {
     cell.type = CellType::kBlock;
   }
 
-  const uint8_t bit = static_cast<uint8_t>((x % 2) * 2 + y % 2);
+  const uint8_t bit = (x % 2) * 2 + y % 2;
   uint8_t value = g_map_block_inversed.at(cell.content.character);
   value |= 1U << bit;
   cell.content.character = g_map_block[value];
@@ -490,7 +490,7 @@ void Canvas::DrawBlockOff(int x, int y) {
   }
   y /= 2;
 
-  const uint8_t bit = static_cast<uint8_t>((y % 2) * 2 + x % 2);
+  const uint8_t bit = (y % 2) * 2 + x % 2;
   uint8_t value = g_map_block_inversed.at(cell.content.character);
   value &= ~(1U << bit);
   cell.content.character = g_map_block[value];
@@ -511,7 +511,7 @@ void Canvas::DrawBlockToggle(int x, int y) {
   }
   y /= 2;
 
-  const uint8_t bit = static_cast<uint8_t>((y % 2) * 2 + x % 2);
+  const uint8_t bit = (y % 2) * 2 + x % 2;
   uint8_t value = g_map_block_inversed.at(cell.content.character);
   value ^= 1U << bit;
   cell.content.character = g_map_block[value];
